+++ sdk/quota/arm-quota/src/azureQuotaExtensionAPI.ts
@@ -54,7 +54,7 @@ export class AzureQuotaExtensionAPI extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-quota/1.0.1`;
+    const packageDetails = `azsdk-js-arm-quota/2.0.0`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`